Changed type data the speed

+++ src/Beings/Animals.java
@@ -2,12 +2,12 @@ package Beings;
 
 abstract class Animals extends Beings{
 
-    int speed;
+    double speed;
     double full = 40;
     boolean live;
     boolean goal;
 
-    Animals(double maxHp, int x, int y, int speed){
+    Animals(double maxHp, int x, int y, double speed){
         super(maxHp, x, y);
         this.speed = speed;
     }
